requirements.txt
pandas<2.1.0
numpy<2.0.0
pydantic==2.12.5
outlines~=1.2.9
PyYAML~=6.0.3
transformers
kernels
accelerate
optuna==4.6.0
scikit-learn==1.7.2
sentencepiece
flair
flair[word-embeddings]
axial-positional-embedding==0.3.12
unsloth